requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.12.10
async-property==0.2.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
dill==0.3.5.1
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
platformdirs==2.5.2
pycares==4.2.2
pycodestyle==2.9.1
pycparser==2.21
pylint==2.15.3
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
